+++ vcd/resource_vcd_vapp_nat_rules.go
@@ -54,6 +54,12 @@ func resourceVcdVappNetworkNatRules() *schema.Resource {
 				ForceNew:    true,
 				Description: "vApp network identifier",
 			},
+			"enabled": &schema.Schema{
+				Type:        schema.TypeBool,
+				Optional:    true,
+				Default:     true,
+				Description: "Enable or disable NAT service. Default is `true`.",
+			},
 			"nat_type": &schema.Schema{
 				Type:         schema.TypeString,
 				Required:     true,
@@ -152,11 +158,17 @@ func resourceVappNetworkNatRulesUpdate(d *schema.ResourceData, meta interface{})
 	if !d.Get("enable_ip_masquerade").(bool) && natType == portForwardingNatType {
 		policy = allowTrafficPolicy
 	}
-	vappNetwork, err := vapp.UpdateNetworkNatRules(networkId, netRules,
+	vappNetwork, err := vapp.UpdateNetworkNatRules(networkId, netRules, d.Get("enabled").(bool),
 		natType, policy)
 	if err != nil {
 		log.Printf("[INFO] Error setting NAT rules: %s", err)
 		return fmt.Errorf("error setting NAT rules: %s", err)
+	}
+
+	if vappNetwork.Configuration.Features.FirewallService != nil &&
+		!vappNetwork.Configuration.Features.FirewallService.IsEnabled &&
+		d.Get("enabled").(bool) {
+		_, _ = fmt.Fprint(getTerraformStdout(), "WARNING: for NAT rules to work, firewall has to be enabled. It can be enabled using vcd_vapp_firewall_rules\n")
 	}
 
 	d.SetId(vappNetwork.ID)
@@ -233,6 +245,7 @@ func resourceVappNetworkNatRulesRead(d *schema.ResourceData, meta interface{}) e
 		}
 		rules = append(rules, singleRule)
 	}
+	_ = d.Set("enabled", vappNetwork.Configuration.Features.NatService.IsEnabled)
 	if vappNetwork.Configuration.Features.NatService.NatType == portForwardingNatType &&
 		vappNetwork.Configuration.Features.NatService.Policy == allowTrafficInPolicy {
 		_ = d.Set("enable_ip_masquerade", true)
@@ -245,6 +258,13 @@ func resourceVappNetworkNatRulesRead(d *schema.ResourceData, meta interface{}) e
 	if err != nil {
 		return err
 	}
+
+	if vappNetwork.Configuration.Features.FirewallService != nil &&
+		!vappNetwork.Configuration.Features.FirewallService.IsEnabled &&
+		d.Get("enabled").(bool) {
+		_, _ = fmt.Fprint(getTerraformStdout(), "WARNING: for NAT rules to work, firewall has to be enabled. It can be enabled using vcd_vapp_firewall_rules\n")
+	}
+
 	return nil
 }