+++ CHANGELOG.md
@@ -10,6 +10,7 @@ FEATURES:
 * **New Resource:** `vcd_org_group` Org Group management [GH-513]
 * **New Resource:** `resource/vcd_vapp_firewall_rules` vApp network firewall rules [GH-511]
 * **New Resource:** `resource/vcd_vapp_nat_rules` vApp network NAT rules [GH-518]
+* **New Resource:** `resource/vcd_vapp_static_routing` vApp network static routing rules [GH-520]
 
 IMPROVEMENTS: